requirements.txt
wheel
pip

aiohttp==3.12.15
aiohappyeyeballs==2.6.1
aioresponses==0.7.8
websockets==15.0.1
yarl==1.20.1

numpy==2.3.2
pandas==2.3.1
pyyaml==6.0.2
regex
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
tqdm==4.67.1
sqlalchemy==2.0.43
orjson==3.11.2
tenacity==9.1.2
pygments==2.19.2
zstandard==0.24.0
loguru==0.7.3
python-dotenv==1.1.1

langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
openai==1.99.9
tiktoken==0.11.0

pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.1

anyio==4.10.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
propcache==0.3.2
requests-toolbelt==1.0.0
pytest==8.4.1
pytest-asyncio==1.1.0

uvicorn
redis
fastapi